# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stash It.",
        "title": "Stash It API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/collections/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete one of the current user's collections by ID. The request must\nsaved_items_action is required and must be \"delete\" or \"move\".\n\"delete\" deletes those saved items along with the collection. \"move\"\ntarget_collection_id must be null when saved_items_action is\nUnsorted itself cannot be deleted. It is identified by its\nThe whole operation is atomic. The saved items are deleted or moved\nThe collection must belong to the authenticated user. A collection\nPossible error codes:\n- BAD_REQUEST\n- INVALID_COLLECTION_DELETE_ACTION\n- INVALID_COLLECTION_DELETE_TARGET\n- INVALID_AUTHORIZATION_HEADER\n- INVALID_ACCESS_TOKEN\n- ACCESS_TOKEN_EXPIRED\n- COLLECTION_NOT_FOUND\n- COLLECTION_DELETE_TARGET_NOT_FOUND\n- UNSORTED_COLLECTION_PROTECTED\n- COLLECTION_NOT_EMPTY\n- INTERNAL_SERVER_ERROR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Delete a collection",
                "parameters": [
                    {
                        "type": "string",
                        "example": "01a0f359-093b-737a-963a-80f7ca6768ed",
                        "description": "Collection ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delete collection request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collection.DeleteCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collection.DeleteCollectionAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid collection ID, action or target",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired access token",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Collection or target collection not found",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Unsorted collection, or collection is not empty",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/saved-items/{id}/collection": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "File one of the current user's saved items into one of their user collections, identified by display name.\nThe collection is created if it does not exist yet. A user collection is never created empty, so this call always ends with the saved item inside it.\nNames are matched case-insensitively and with surrounding whitespace ignored, and the stored name is trimmed.\nThis operation is idempotent. If the saved item is already in the target collection nothing is written, the saved item's updated_at is left unchanged, and already_in_collection is returned true.\nA name already held by a system collection, such as \"Unsorted\", is reserved. It returns COLLECTION_NAME_RESERVED rather than filing the saved item into a system collection.\nThe saved item must belong to the authenticated user. A saved item that does not exist and one owned by another user both return the same not found response, so this endpoint never discloses whether an ID exists.\nPossible error codes:\n- INVALID_COLLECTION_NAME\n- COLLECTION_NAME_RESERVED\n- BAD_REQUEST\n- SAVED_ITEM_NOT_FOUND\n- INVALID_AUTHORIZATION_HEADER\n- INVALID_ACCESS_TOKEN\n- ACCESS_TOKEN_EXPIRED\n- INTERNAL_SERVER_ERROR",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collection"
                ],
                "summary": "Put a saved item into a collection",
                "parameters": [
                    {
                        "type": "string",
                        "example": "01a0f359-093b-737a-963a-80f7ca6768ed",
                        "description": "Saved item ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Put saved item into collection request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/collection.PutSavedItemIntoCollectionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/collection.PutSavedItemIntoCollectionAPIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid collection name or saved item ID",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid or expired access token",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Saved item not found",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Collection name is reserved",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/swagger.APIErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "collection.CollectionResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2026-10-02T10:30:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "01a0f359-093b-737a-963a-80f7ca6768ed"
                },
                "name": {
                    "type": "string",
                    "example": "Wishlist"
                },
                "type": {
                    "type": "string",
                    "example": "user"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-10-02T10:30:00Z"
                }
            }
        },
        "collection.DeleteCollectionAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "collection deleted successfully"
                }
            }
        },
        "collection.DeleteCollectionRequest": {
            "type": "object",
            "properties": {
                "saved_items_action": {
                    "description": "SavedItemsAction is what to do with the collection's saved items.\n\n\"delete\" removes them along with the collection. \"move\" files them into\ntarget_collection_id first, and then removes the collection. Both are\nrequired: the request must say which, because the two are not variants of a\nsingle outcome and one of them discards saved items.",
                    "enum": [
                        "delete",
                        "move"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/collection.SavedItemsAction"
                        }
                    ],
                    "example": "delete"
                },
                "target_collection_id": {
                    "description": "TargetCollectionID is where the saved items go when SavedItemsAction is\n\"move\".\n\nIt must be a collection id and must be null when SavedItemsAction is\n\"delete\". Unsorted is a valid target and is named here by its id, like any\nother collection; there is no separate mode or flag for it, and no fallback\nto it. A target that does not exist is an error rather than a substitution.",
                    "type": "string",
                    "example": "01a0f359-093b-737a-963a-80f7ca6768ed"
                }
            }
        },
        "collection.PutSavedItemIntoCollectionAPIResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/collection.PutSavedItemIntoCollectionResponse"
                },
                "message": {
                    "type": "string",
                    "example": "saved item moved into collection successfully"
                }
            }
        },
        "collection.PutSavedItemIntoCollectionRequest": {
            "type": "object",
            "properties": {
                "collection_name": {
                    "description": "CollectionName is the display name of the user collection to file the saved\nitem under. The collection is created if it does not exist yet. Names are\nmatched case-insensitively with surrounding whitespace ignored, and a name\nalready used by a system collection is rejected.\n\nThis field carries no `validate` constraint on purpose. The service owns\ncollection name rules and answers with the INVALID_COLLECTION_NAME code, so\na constraint here would pre-empt that and collapse both failure cases into a\ngeneric VALIDATION_ERROR.",
                    "type": "string",
                    "example": "Wishlist"
                }
            }
        },
        "collection.PutSavedItemIntoCollectionResponse": {
            "type": "object",
            "properties": {
                "already_in_collection": {
                    "description": "AlreadyInCollection is true when the saved item was already in the target\ncollection, in which case nothing was written. It is mutually exclusive with\nCollectionCreated.",
                    "type": "boolean",
                    "example": false
                },
                "collection": {
                    "$ref": "#/definitions/collection.CollectionResponse"
                },
                "collection_created": {
                    "description": "CollectionCreated is true only when this call created the collection. It is\nfalse when an existing collection was reused.",
                    "type": "boolean",
                    "example": true
                },
                "saved_item": {
                    "$ref": "#/definitions/collection.PutSavedItemSavedItemResponse"
                }
            }
        },
        "collection.PutSavedItemSavedItemResponse": {
            "type": "object",
            "properties": {
                "collection_id": {
                    "type": "string",
                    "example": "01a0f359-093b-737a-963a-80f7ca6768ed"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-10-02T10:30:00Z"
                },
                "domain": {
                    "type": "string",
                    "example": "example.com"
                },
                "id": {
                    "type": "string",
                    "example": "01a0f359-093b-737a-963a-80f7ca6768ed"
                },
                "platform": {
                    "type": "string",
                    "example": "youtube"
                },
                "title": {
                    "type": "string",
                    "example": "An interesting article"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2026-10-02T10:30:00Z"
                },
                "url": {
                    "type": "string",
                    "example": "https://example.com/articles/1"
                }
            }
        },
        "collection.SavedItemsAction": {
            "type": "string",
            "enum": [
                "delete",
                "move"
            ],
            "x-enum-varnames": [
                "SavedItemsActionDelete",
                "SavedItemsActionMove"
            ]
        },
...
        "swagger.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "ERROR_CODE"
                },
                "message": {
                    "type": "string",
                    "example": "request failed"
                }
            }
        },
...
        "swagger.APIErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/swagger.APIError"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your access token using the Bearer scheme. Example: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}